Move plan and sim result implementations to remote package
