+++ e2e-tests/src/test/java/gov/nasa/jpl/aerie/e2e/TimelineRemoteTests.java
@@ -10,8 +10,8 @@ import gov.nasa.jpl.aerie.timeline.Interval;
 import gov.nasa.jpl.aerie.timeline.collections.profiles.Real;
 import gov.nasa.jpl.aerie.timeline.payloads.LinearEquation;
 import gov.nasa.jpl.aerie.timeline.payloads.Segment;
-import gov.nasa.jpl.aerie.timeline.plan.AeriePostgresPlan;
-import gov.nasa.jpl.aerie.timeline.plan.AeriePostgresSimulationResults;
+import gov.nasa.jpl.aerie.procedural.remote.AeriePostgresPlan;
+import gov.nasa.jpl.aerie.procedural.remote.AeriePostgresSimulationResults;
 import gov.nasa.jpl.aerie.timeline.plan.SimulatedPlan;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
