+++ procedural/remote/MODULE_DOCS.md
@@ -1,0 +1,8 @@
+# Module Remote
+
+This library provides tools for accessing plans and simulation results from remote execution environments.
+This is intended to allow goal and constraint authors to run their code outside Aerie, and these implementations
+do not need to be packaged in the goal and constraint jars.
+
+# Package gov.nasa.jpl.aerie.procedural.remote
+Remote utilities for running goals and constraints outside of Aerie.
