+++ procedural/remote/build.gradle
@@ -13,8 +13,11 @@ repositories {
 }
 
 dependencies {
+  implementation project(':procedural:scheduling')
   implementation project(':procedural:timeline')
   implementation project(':merlin-driver')
+  implementation project(':scheduler-driver')
+  implementation project(':parsing-utilities')
 
   testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
   testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
@@ -34,12 +37,13 @@ kotlin {
 }
 
 var timelineSource = "${project(":procedural:timeline").projectDir}/src/main/kotlin"
+var schedulingSource = "${project(":procedural:scheduling").projectDir}/src/main/kotlin"
 
 dokkaHtmlPartial.configure {
   dokkaSourceSets {
     configureEach {
       // used as project name in the header
-      moduleName.set("Scheduling")
+      moduleName.set("Remote")
 
       reportUndocumented.set(true)
 
@@ -48,7 +52,8 @@ dokkaHtmlPartial.configure {
 
       sourceRoots.from(timelineSource)
       suppressedFiles.from(timelineSource)
+      sourceRoots.from(schedulingSource)
+      suppressedFiles.from(schedulingSource)
     }
   }
 }
-
