+++ procedural/remote/src/main/kotlin/gov/nasa/jpl/aerie/procedural/remote/SqlUtils.kt
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.timeline.plan
+package gov.nasa.jpl.aerie.procedural.remote
 
 import java.sql.Connection
 import java.sql.PreparedStatement
