+++ procedural/timeline/src/main/kotlin/gov/nasa/jpl/aerie/timeline/payloads/activities/AnyInstance.kt
@@ -1,7 +1,6 @@
 package gov.nasa.jpl.aerie.timeline.payloads.activities
 
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue
-import gov.nasa.jpl.aerie.timeline.plan.AeriePostgresPlan
 import kotlin.jvm.optionals.getOrNull
 
 /** A general-purpose container for representing the arguments and computed attributes of any type of activity instance. */
@@ -14,10 +13,12 @@ data class AnyInstance(
      * Converts a [SerializedValue] object containing activity arguments and computed attributes to an [AnyInstance] object.
      */
     fun deserialize(attributes: SerializedValue): AnyInstance {
+      /***/ class InstanceDeserializeError(message: String): Error(message)
+
       val arguments = attributes.asMap().getOrNull()!!["arguments"]?.asMap()?.getOrNull()
-          ?: throw AeriePostgresPlan.DatabaseError("Could not get arguments from attributes: $attributes")
+          ?: throw InstanceDeserializeError("Could not get arguments from attributes: $attributes")
       val computedAttributes = attributes.asMap().getOrNull()!!["computedAttributes"]
-          ?: throw AeriePostgresPlan.DatabaseError("Could not get computed attributes from attributes: $attributes")
+          ?: throw InstanceDeserializeError("Could not get computed attributes from attributes: $attributes")
       return AnyInstance(arguments, computedAttributes)
     }
   }
